+++ exonum/src/api/mod.rs
@@ -29,6 +29,7 @@ pub mod backends;
 pub mod error;
 pub mod node;
 mod state;
+pub(crate) mod websocket;
 mod with;
 
 /// Defines object that could be used as an API backend.
@@ -274,7 +275,10 @@ impl ApiAggregator {
             "system".to_owned(),
             Self::system_api(&blockchain, node_state.clone()),
         );
-        inner.insert("explorer".to_owned(), Self::explorer_api());
+        inner.insert(
+            "explorer".to_owned(),
+            Self::explorer_api(node_state.clone()),
+        );
         // Adds services APIs.
         inner.extend(blockchain.service_map().iter().map(|(_, service)| {
             let mut builder = ServiceApiBuilder::new();
@@ -317,9 +321,9 @@ impl ApiAggregator {
         self.inner.insert(prefix.into(), builder);
     }
 
-    fn explorer_api() -> ServiceApiBuilder {
+    fn explorer_api(shared_node_state: SharedNodeState) -> ServiceApiBuilder {
         let mut builder = ServiceApiBuilder::new();
-        self::node::public::ExplorerApi::wire(builder.public_scope());
+        self::node::public::ExplorerApi::wire(builder.public_scope(), shared_node_state);
         builder
     }