+++ exonum/src/api/node/public/explorer.rs
@@ -14,12 +14,20 @@
 
 //! Exonum blockchain explorer API.
 
+use actix::Arbiter;
+use actix_web::{http, ws};
+use futures::IntoFuture;
 use serde_json;
 
 use std::ops::Range;
+use std::sync::{Arc, Mutex};
 
-use api::{Error as ApiError, ServiceApiScope, ServiceApiState};
-use blockchain::Block;
+use api::{
+    backends::actix::{self, FutureResponse, HttpRequest, RawHandler, RequestHandler},
+    websocket::{Server, Session}, Error as ApiError, ServiceApiBackend, ServiceApiScope,
+    ServiceApiState,
+};
+use blockchain::{Block, SharedNodeState};
 use crypto::Hash;
 use explorer::{BlockchainExplorer, TransactionInfo};
 use helpers::Height;
@@ -156,8 +164,45 @@ impl ExplorerApi {
             })
     }
 
+    /// Subscribes to block commits events.
+    pub fn handle_subscribe(
+        name: &'static str,
+        backend: &mut actix::ApiBuilder,
+        shared_node_state: SharedNodeState,
+    ) {
+        let server = Arc::new(Mutex::new(None));
+
+        let index = move |req: HttpRequest| -> FutureResponse {
+            let server = server.clone();
+            let mut address = server.lock().expect("Expected mutex lock");
+            if address.is_none() {
+                *address = Some(Arbiter::start(|_| Server::default()));
+
+                shared_node_state.set_broadcast_server_address(address.to_owned().unwrap());
+            }
+
+            Box::new(
+                ws::start(req.clone(), Session::new(address.to_owned().unwrap())).into_future(),
+            )
+        };
+
+        backend.raw_handler(RequestHandler {
+            name: name.to_owned(),
+            method: http::Method::GET,
+            inner: Arc::from(index) as Arc<RawHandler>,
+        });
+    }
+
     /// Adds explorer API endpoints to the corresponding scope.
-    pub fn wire(api_scope: &mut ServiceApiScope) -> &mut ServiceApiScope {
+    pub fn wire(
+        api_scope: &mut ServiceApiScope,
+        shared_node_state: SharedNodeState,
+    ) -> &mut ServiceApiScope {
+        Self::handle_subscribe(
+            "v1/blocks/subscribe",
+            api_scope.web_backend(),
+            shared_node_state,
+        );
         api_scope
             .endpoint("v1/blocks", Self::blocks)
             .endpoint("v1/block", Self::block)