+++ CHANGELOG.md
@@ -14,6 +14,13 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
 - `NodePublicConfig` `addr` field has been renamed to `address` (#809).
 
+### New Features
+
+#### exonum
+
+- Added `/v1/blocks/subscribe` endpoint for following block commit events
+  through WebSockets (#792).
+
 ### Internal Improvements
 
 - Transactions (signature) verification benchmark has been added. (#808)